+++ MyGame/MapLayer.cpp
@@ -13,7 +13,7 @@ CMapLayer::CMapLayer(TiXmlElement* layer, float width, float height) {
     //chuyển string vector thành int để vẽ
     vector<string> tileIdSetRaw = split(mapInfo, ",");
 
-    int id =0;
+    size_t id =0;
     
     for (int i = 0; i < height; i++)
     {
@@ -32,10 +32,10 @@ void CMapLayer::SetRenderRegion(int& cell_startX, int& cell_startY, int& cell_en
     float camX, camY;
     CGame::GetInstance()->GetMainCamera()->GetPosition(camX, camY);
 
-    cell_startX = camX / TILE_WIDTH;
-    cell_startY = camY / TILE_WIDTH;
-    cell_endX = (camX + CAM_WIDTH) / TILE_WIDTH ;
-    cell_endY = (camY + CAM_HEIGHT) / TILE_WIDTH ;
+    cell_startX = (int)camX / TILE_WIDTH;
+    cell_startY = (int)camY / TILE_WIDTH;
+    cell_endX = (int)(camX + CAM_WIDTH) / TILE_WIDTH ;
+    cell_endY = (int)(camY + CAM_HEIGHT) / TILE_WIDTH ;
 }
